# Conflicts:
#	integration-tests/src/test/java/cc/blynk/integration/http/HttpAPIPinsAsyncClientTest.java
#	integration-tests/src/test/java/cc/blynk/integration/tcp/DeviceTilesWidgetTest.java
#	integration-tests/src/test/java/cc/blynk/integration/tcp/SyncWorkflowTest.java
#	server/core/src/main/java/cc/blynk/server/core/dao/ReportingDiskDao.java
#	server/core/src/main/java/cc/blynk/server/core/model/DashBoard.java
#	server/core/src/main/java/cc/blynk/server/core/model/DataStream.java
#	server/core/src/main/java/cc/blynk/server/core/reporting/average/AggregationKey.java
#	server/core/src/main/java/cc/blynk/server/core/reporting/raw/BaseReportingKey.java
#	server/core/src/main/java/cc/blynk/server/db/dao/ReportingDBDao.java
#	server/core/src/test/java/cc/blynk/server/core/model/DataStreamValuesUpdateCorrectTest.java
#	server/core/src/test/java/cc/blynk/server/core/model/widgets/DataStreamGetJsonValueTest.java
#	server/core/src/test/java/cc/blynk/server/core/model/widgets/MultiPinWidgetsToJsonTest.java
#	server/core/src/test/java/cc/blynk/server/db/RawDataDBTest.java
#	server/core/src/test/java/cc/blynk/server/db/RealtimeStatsDBTest.java
#	server/launcher/src/test/java/cc/blynk/server/workers/ReportingWorkerTest.java
#	server/tcp-app-server/src/main/java/cc/blynk/server/application/handlers/main/logic/MobileHardwareResendFromBTLogic.java

+++ integration-tests/src/test/java/cc/blynk/integration/tcp/SyncWorkflowTest.java
@@ -328,7 +328,7 @@ public class SyncWorkflowTest extends SingleServerInstancePerTest {
         clientPair.appClient.reset();
         clientPair.appClient.send("loadProfileGzipped");
         Profile profile = clientPair.appClient.parseProfile(1);
-        TimeInput timeInput = (TimeInput) profile.dashBoards[0].findWidgetByPin(0, (byte) 99, PinType.VIRTUAL);
+        TimeInput timeInput = (TimeInput) profile.dashBoards[0].findWidgetByPin(0, (short) 99, PinType.VIRTUAL);
         assertNotNull(timeInput);
         assertEquals(82800, timeInput.startAt);
         assertEquals(82860, timeInput.stopAt);
