+++ server/core/src/main/java/cc/blynk/server/core/dao/ReportingDiskDao.java
@@ -90,7 +90,7 @@ public class ReportingDiskDao implements Closeable {
     }
 
     public ByteBuffer getByteBufferFromDisk(User user, int dashId, int deviceId,
-                                            PinType pinType, byte pin, int count,
+                                            PinType pinType, short pin, int count,
                                             GraphGranularityType type, int skipCount) {
         Path userDataFile = Paths.get(
                 dataFolder,
@@ -216,7 +216,7 @@ public class ReportingDiskDao implements Closeable {
         return false;
     }
 
-    private static String generateFilename(int dashId, int deviceId, char pinType, byte pin, String type) {
+    private static String generateFilename(int dashId, int deviceId, char pinType, short pin, String type) {
         return generateFilenamePrefix(dashId, deviceId) + pinType + pin + "_" + type + ".bin";
     }
 
@@ -228,7 +228,7 @@ public class ReportingDiskDao implements Closeable {
         return "history_" + dashId + DEVICE_SEPARATOR + deviceId + "_";
     }
 
-    private static void delete(String userReportingDir, int dashId, int deviceId, PinType pinType, byte pin,
+    private static void delete(String userReportingDir, int dashId, int deviceId, PinType pinType, short pin,
                                GraphGranularityType reportGranularity) {
         Path userDataFile = Paths.get(userReportingDir,
                 generateFilename(dashId, deviceId, pinType, pin, reportGranularity));
@@ -236,7 +236,7 @@ public class ReportingDiskDao implements Closeable {
     }
 
     public static String generateFilename(int dashId, int deviceId,
-                                          PinType pinType, byte pin, GraphGranularityType type) {
+                                          PinType pinType, short pin, GraphGranularityType type) {
         return generateFilename(dashId, deviceId, pinType.pintTypeChar, pin, type.label);
     }
 
@@ -280,7 +280,7 @@ public class ReportingDiskDao implements Closeable {
         return count;
     }
 
-    public void delete(User user, int dashId, int deviceId, PinType pinType, byte pin) {
+    public void delete(User user, int dashId, int deviceId, PinType pinType, short pin) {
         log.debug("Removing {}{} pin data for dashId {}, deviceId {}.", pinType.pintTypeChar, pin, dashId, deviceId);
         String userReportingDir = getUserReportingFolderPath(user).toString();
 
@@ -289,7 +289,7 @@ public class ReportingDiskDao implements Closeable {
         }
     }
 
-    public void process(User user, DashBoard dash, Device device, byte pin, PinType pinType, String value, long ts) {
+    public void process(User user, DashBoard dash, Device device, short pin, PinType pinType, String value, long ts) {
         try {
             double doubleVal = NumberUtil.parseDouble(value);
             process(user, dash, device, pin, pinType, value, ts, doubleVal);
@@ -299,7 +299,7 @@ public class ReportingDiskDao implements Closeable {
         }
     }
 
-    private void process(User user, DashBoard dash, Device device, byte pin, PinType pinType,
+    private void process(User user, DashBoard dash, Device device, short pin, PinType pinType,
                          String value, long ts, double doubleVal) {
         int deviceId = device.id;
         if (enableRawDbDataStore) {
