+++ server/core/src/test/java/cc/blynk/server/core/reporting/average/AverageAggregatorTest.java
@@ -40,7 +40,7 @@ public class AverageAggregatorTest {
 
         PinType pinType = PinType.VIRTUAL;
         int dashId = 1;
-        byte pin = 1;
+        short pin = 1;
 
         long ts = getMillis(2015, 8, 1, 0, 0);
 
@@ -67,7 +67,7 @@ public class AverageAggregatorTest {
         user.email = "test@test.com";
         PinType pinType = PinType.VIRTUAL;
         int dashId = 1;
-        byte pin = 1;
+        short pin = 1;
 
         double expectedDailyAverage = 0;
 
@@ -108,7 +108,7 @@ public class AverageAggregatorTest {
         user.email = "test@test.com";
         PinType pinType = PinType.VIRTUAL;
         int dashId = 1;
-        byte pin = 1;
+        short pin = 1;
 
         double expectedDailyAverage = 0;
